+++ src-be/src/services/shopping-list-service.ts
@@ -1,7 +1,13 @@
 import { FilterQuery } from 'mongoose';
 import { IShoppingList, ShoppingList } from '../models/shopping-list.js';
+import { AuthError } from '../models/errors/auth-error.js';
 
 class ShoppingListService {
+    async exists(filters: FilterQuery<IShoppingList>): Promise<boolean> {
+        const exists = await ShoppingList.exists(filters);
+        return exists != null;
+    }
+
     async find(filters: FilterQuery<IShoppingList>): Promise<IShoppingList[]> {
         return await ShoppingList.find(filters);
     }
@@ -10,7 +16,14 @@ class ShoppingListService {
         await ShoppingList.create(model);
     }
 
-    async delete(id: string) {
+    async isOwner(userId: string, shoppingListId: string): Promise<boolean> {
+        return await this.exists({ _id: shoppingListId, idOwner: userId });
+    }
+
+    async delete(id: string, userId: string) {
+        if (!await this.isOwner(userId, id)) 
+            throw new AuthError("Only owner can do this!");
+
         await ShoppingList.findByIdAndDelete(id);
     }
 }
